// src/shared/api/generated/model/sessionRecordsResponseDtoTotalsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */

export type SessionRecordsResponseDtoTotalsItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  staffMembershipId: string;
  staffName: string;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  classCount: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  plannedSeconds: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  actualSeconds: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  openCount: number;
};